Keep every legacy test verdict on the line that names its test (#40)
Two tests in src/tests.cpp printed " --- Success" on a line of its own:
testRemovingMicroorganismFromMicrobiome ended its name with std::endl,
and testWebServerServesSimulationState printed its name before starting
a WebServer whose run() writes its own "Serving ..." line. The first
drops the endl; the second prints its name only once the server has
been stopped and joined, so the verdict follows the name directly.

Also make the simulation loop counter in mbapp.cpp a size_t to match
AppConfig::getNumSimulations(), removing the only -Wsign-compare
warning in the build.

Closes #38
Closes #39

+++ src/mbapp.cpp
@@ -7,7 +7,7 @@ MicrobiomeApplication::MicrobiomeApplication() {
 
 bool MicrobiomeApplication::run() {
     // run simulations
-    for (int i = 0; i < config.getNumSimulations(); i++) {
+    for (size_t i = 0; i < config.getNumSimulations(); i++) {
         std::cout << "Running simulation " << i + 1 << " of " << config.getNumSimulations() << std::endl;
         simulation = new Simulation(&config, i + 1, "Simulation " + std::to_string(i + 1));
         simulation->run();
